refactor(session): 사주·신점 세션 페이지 컨트롤러 훅 분리 (B-5 잔여)
#437(타로·설정)에 이은 B-5 마무리. 동일 패턴 — 로직을 훅으로 verbatim 이동해
동작·DOM·E2E 셀렉터(data-testid 포함) 보존.

- 사주 세션: 리딩진행 로직(SSE·대기연출·240s 타임아웃) → useSajuReading. page 331→160줄
- 신점 세션: 대화·최종리딩 로직 → useShinjeomChat. page 411→178줄
- 신규 훅 2개 sonar.coverage/cpd.exclusions 동기화, CLAUDE.md 훅 목록 갱신
- trivial 동작불변 정리: saju replaceAll·Number.parseInt (verbatim 범위 내)

검증: type-check·lint(0 errors, 신규 훅 경고 0)·build 성공. DOM/JSX 불변=E2E 셀렉터 영향 0.
이로써 세션 4개 페이지(타로·사주·신점·설정) 전부 <300줄, view/controller 분리 완료.
