+++ sonar-project.properties
@@ -58,6 +58,8 @@ sonar.coverage.exclusions=\
   src/hooks/useTarotReading.ts,\
   src/hooks/useTarotCardSelection.ts,\
   src/hooks/useSettings.ts,\
+  src/hooks/useSajuReading.ts,\
+  src/hooks/useShinjeomChat.ts,\
   src/hooks/useMouseParallax.ts,\
   src/components/effects/ServiceIllustrations.tsx,\
   src/lib/supabase/client.ts,\
@@ -117,6 +119,8 @@ sonar.cpd.exclusions=\
   src/hooks/useTarotReading.ts,\
   src/hooks/useTarotCardSelection.ts,\
   src/hooks/useSettings.ts,\
+  src/hooks/useSajuReading.ts,\
+  src/hooks/useShinjeomChat.ts,\
   src/components/saju/SajuChartReveal.tsx,\
   src/components/shinjeom/ShinjeomEnergyEffect.tsx,\
   src/components/session/ResultTextCard.tsx,\
